Found a Guice-related bug in ConsoleLogFilterStep.
